jeju_local_spots/requirements.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
pillow==11.0.0
sqlparse==0.5.3
tzdata==2024.2
